+++ apps/gi-backend/src/app/weapon/weapon.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common'
 import type { Weapon } from '@prisma/client/gi'
+import { GraphQLError } from 'graphql'
 import { PrismaService } from '../prisma/prisma.service'
+import type { UpdateWeapon } from './weapon.entity'
 
 @Injectable()
 export class WeaponService {
@@ -26,7 +28,38 @@ export class WeaponService {
     return this.prisma.weapon.findMany({ where: { genshinUserId } })
   }
 
-  remove(id: string) {
-    return this.prisma.weapon.delete({ where: { id } })
+  remove(id: string, genshinUserId: string) {
+    return this.prisma.weapon.delete({ where: { id, genshinUserId } })
+  }
+  async update(weapon: UpdateWeapon, genshinUserId: string) {
+    if (weapon.location) {
+      // TODO: this doesnt validate weapon type vs character weapon type.
+      const oldWeapon = await this.prisma.weapon.findUnique({
+        where: { id: weapon.id },
+      })
+      if (!oldWeapon) throw new GraphQLError('Weapon does not exist')
+      if (oldWeapon.location !== weapon.location && weapon.location) {
+        // update other weapon that could be equipped to the same char
+        const where = {
+          genshinUserId_location: {
+            genshinUserId,
+            location: weapon.location,
+          },
+        }
+        const conflictWeapon = await this.prisma.weapon.findUnique({ where })
+        if (conflictWeapon)
+          await this.prisma.weapon.update({
+            where,
+            data: {
+              location: oldWeapon.location,
+            },
+          })
+      }
+    }
+    const { id, ...data } = weapon
+    return this.prisma.weapon.update({
+      where: { id, genshinUserId },
+      data,
+    })
   }
 }
